Nabin166

+++ BankFull/Models/TransactionRate.cs
@@ -6,7 +6,7 @@ namespace BankFull.Models
     {
         [Key]
         public int Id { get; set; }
-        public float Rate { get; set; }
+        public decimal Rate { get; set; }
        
         
         public string Date { get; set; }
